backend/vector_store.py
from langchain.vectorstores import S3VectorStore
from langchain.embeddings.openai import OpenAIEmbeddings
import os
from dotenv import load_dotenv

load_dotenv()

embeddings = OpenAIEmbeddings(openai_api_key=os.getenv('OPENAI_API_KEY'))

vector_store = S3VectorStore(
    bucket_name=os.getenv('AWS_S3_BUCKET'),
    embedding_function=embeddings,
    region_name=os.getenv('AWS_REGION'),
    aws_access_key_id=os.getenv('AWS_ACCESS_KEY_ID'),
    aws_secret_access_key=os.getenv('AWS_SECRET_ACCESS_KEY')
)
