backend/requirements.txt
flask
python-dotenv
boto3
langchain
openai
